Fix bfp16 regressions on conv based models (#2105)
### Summary 

- This PR fixes [issue
#2095](https://github.com/tenstorrent/tt-forge-fe/issues/2095) and
addresses feedback from [PR
#2098](https://github.com/tenstorrent/tt-forge-fe/pull/2098#pullrequestreview-2857843338).
- Testing was performed on the [regressions reported
here](https://github.com/tenstorrent/tt-forge-fe/issues/1987#issuecomment-2888227666)
using [this fix](https://github.com/tenstorrent/tt-forge-fe/pull/2076),
and this PR adds the corresponding changes based on the results.

### Logs


[may22_alexnet.log](https://github.com/user-attachments/files/20394770/may22_alex_th.log)

[may22_deit.log](https://github.com/user-attachments/files/20394771/may22_deit.log)

[may22_densenet.log](https://github.com/user-attachments/files/20394774/may22_densenet.log)

[may22_densenet_121.log](https://github.com/user-attachments/files/20394773/may22_densenet_121.log)

[may22_glpn_kitti_1.log](https://github.com/user-attachments/files/20394819/may22_glpn_kitti_1.log)

[may22_mgp.log](https://github.com/user-attachments/files/20394776/may22_mgp_af.log)

[may22_mlp_timm.log](https://github.com/user-attachments/files/20394845/may22_mlp_timm.log)

[may22_perciever_1.log](https://github.com/user-attachments/files/20394858/may22_perciever_1.log)

[may22_perciever.log](https://github.com/user-attachments/files/20394860/may22_perciever.log)

[may22_rcnn.log](https://github.com/user-attachments/files/20394876/may22_rcnn.log)

[may22_swin_torchvision.log](https://github.com/user-attachments/files/20394777/may22_swin_tv_af.log)

[may22_swin_torchvision_1.log](https://github.com/user-attachments/files/20394779/may22_swin_tv_b.log)

[may22_vgg_timm_hf_torchbhub.log](https://github.com/user-attachments/files/20394780/may22_vgg_timm_th_timm.log)

[may22_vgg_tv.log](https://github.com/user-attachments/files/20394781/may22_vgg_tv.log)

[may22_vit_hf.log](https://github.com/user-attachments/files/20394782/may22_vit_hf.log)

[may22_vit_tv_af.log](https://github.com/user-attachments/files/20394783/may22_vit_tv_af.log)

+++ forge/test/models/pytorch/vision/mnist/model_utils/utils.py
@@ -1,7 +1,6 @@
 # SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
 #
 # SPDX-License-Identifier: Apache-2.0
-import torch
 from torch.utils.data import DataLoader
 from torchvision import datasets, transforms
 
@@ -19,4 +18,4 @@ def load_input():
     test_dataset = datasets.MNIST(root="./data", train=False, transform=transform, download=True)
     dataloader = DataLoader(test_dataset, batch_size=1)
     test_input, _ = next(iter(dataloader))
-    return [test_input.to(torch.bfloat16)]
+    return [test_input]
